fix(employee-work-details): expected_hours auf f64 statt i64 binden im DAO-update
`EmployeeWorkDetailsDaoImpl::update` hat `entity.expected_hours` (f32) als `i64` an
sqlx gebunden — dadurch wurde z.B. 40.25 beim Speichern eines bestehenden
Arbeitsvertrags auf 40 truncated. Der `create()`-Pfad war von Anfang an korrekt
mit `as f64`; der Bug fiel erst auf, nachdem Commit 72e045f dem Frontend `step=0.01`
gegeben hat und damit Dezimal-Eingaben überhaupt erst die Validation passieren.

+3 Roundtrip-Regression-Tests (Service → DAO → in-memory SQLite → DAO → Service)
für create/update mit fractional und 2-decimal expected_hours, sowie die zugehörige
Debug-Session-Doku unter .planning/debug/working-hours-truncated.md.

+++ dao_impl_sqlite/src/employee_work_details.rs
@@ -413,7 +413,7 @@ impl EmployeeWorkDetailsDao for EmployeeWorkDetailsDaoImpl {
         let to_year = entity.to_year as i64;
         let to_calendar_week = entity.to_calendar_week as i64;
         let to_day_of_week = entity.to_day_of_week.to_number() as i64;
-        let expected_hours = entity.expected_hours as i64;
+        let expected_hours = entity.expected_hours as f64;
         let vacation_days = entity.vacation_days as i64;
         let workdays_per_week = entity.workdays_per_week as i64;
         let is_dynamic = entity.is_dynamic as i64;
